pyproject.toml
[project]
name = "datex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "minuth", email = "marvin.minuth@jam-software.de" }
]
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.5.1",
    "openai>=1.93.0",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.1.1",
]

[project.scripts]
datex = "datex:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
